♻️ Refactor metrics
