+++ .gitignore
@@ -9,8 +9,6 @@ __pycache__/
 # C extensions
 *.so
 
-*.pth
-
 # Distribution / packaging
 .Python
 build/
