+++ utils/metrics/__init__.py
@@ -1,0 +1,2 @@
+from utils.metrics.lpips.lpips import calculate_lpips_given_paths
+from utils.metrics.pytorch_fid.fid_score import calculate_fid_given_paths
